requirements.txt
aioredis==1.2.0
alabaster==0.7.12
asgiref==3.2.1
asn1crypto==0.24.0
astroid==1.6.6
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
Babel==2.6.0
beautifulsoup4==4.8.0
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect2==0.2.1
daphne==2.3.0
defusedxml==0.6.0
Django==2.0.1
django-angular-dynamic-forms==6.1.2
django-common-helpers==0.9.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.0.0
django-rq==1.0.1
djangorestframework==3.8.2
dnspython==1.16.0
docutils==0.14
email-validator==1.0.4
gunicorn==19.9.0
hiredis==1.0.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.6
imagesize==1.0.0
incremental==17.5.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
msgpack==0.6.1
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.12
num2words==0.5.6
packaging==17.1
pdfrw==0.4
phonenumbers==8.10.15
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.3.1
pycparser==2.18
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==1.8.2
pylint-django==0.9.4
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
Pyphen==0.9.4
python-dateutil==2.8.0
pytz==2018.3
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
rq==0.10.0
six==1.11.0
snowballstemmer==1.2.1
soupsieve==1.9.2
Sphinx==1.7.2
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
tinycss2==0.6.1
Twisted==19.7.0
txaio==18.8.1
typing==3.6.4
urllib3==1.22
WeasyPrint==0.42.2
webencodings==0.5.1
wrapt==1.10.11
xlrd==1.2.0
zope.interface==4.6.0
